+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -7,6 +7,7 @@ Release History
 ++++++
 * `vm extension show / wait`: deprecated --expand parameter.
 * `vm restart`: Added `--force` which redeploys unresponsive VMs.
+* `vm/vmss create`: `--authentication-type` now accepts/infers "all" to create a VM with both password and ssh authentication.
 
 2.2.8
 ++++++